backend/requirements.txt
fastapi
uvicorn
sentence-transformers
nltk
scikit-learn
numpy
python-multipart
requests
sqlalchemy
pymysql
python-dotenv
cryptography
